refactor: move bookservice to examples/protogen/bookservice
Moved from ext/protogen/examples/ to top-level examples/ alongside
the apps examples. Updated go.mod replace paths, proto go_package,
regenerated proto output, and updated protogen Makefile test-e2e path.

Structure:
  examples/
  ├── apps/{vanilla,htmx,react}
  └── protogen/bookservice
