+++ examples/protogen/bookservice/Makefile
