+++ examples/protogen/bookservice/buf.gen.yaml
