+++ examples/protogen/bookservice/buf.lock
@@ -1,6 +1,9 @@
 # Generated by buf. DO NOT EDIT.
 version: v2
 deps:
+  - name: buf.build/mcpkit/protogen
+    commit: 2a6292c4ec9f46c9ab4236b328ae33ef
+    digest: b5:daa88f609479996ed190b6be15f666331c952d97ba387f5a316b2fb14e8c3b2bb15d58e1a3123d5a0b173e83ebcdf971db2d80768a513447ecebcf4d5c567b53
   - name: buf.build/protocolbuffers/wellknowntypes
     commit: 9d16d599a978406980f6e2f081331a93
     digest: b5:dd06e497a5c52f5ddf6ec02b3c7d289cc6c0432093fc2f6bf7a4fb5fae786c3e4c893e55d2759ffb6833268daf3de0bce303a406fed15725790528f2c27dc219
