+++ examples/protogen/bookservice/gen/bookservice/v1/service.pb.mcp.go
