+++ examples/protogen/bookservice/proto/mcp
